+++ 2026/BS/tools/shared/src/Funcs.kt
@@ -1,5 +1,0 @@
-import io.ktor.client.HttpClient
-
-expect fun getEnv(key: String): String
-expect fun createClient(): HttpClient
-expect fun getCurrentTimestamp(): String