+++ 2026/BS/tools/shared/src@linux/Funcs.kt
@@ -8,5 +8,3 @@ import libcurl.curl_getenv
 @OptIn(ExperimentalForeignApi::class)
 actual fun getEnv(key: String): String = curl_getenv(key)?.toKString() ?: ""
 actual fun createClient() = HttpClient(Curl)
-@OptIn(ExperimentalForeignApi::class)
-actual fun getCurrentTimestamp(): String = platform.posix.time(null).toString()